+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+1.3.2 (2021-08-30)
+------------------
+
+- starlette integration added
+- django integration added
+
+
 1.3.1 (2021-08-24)
 ------------------